[refractor] 일기, 답장, 발송기록 테이블 id 타입 Long으로 변경 / [feat] 일기 자동 발송 기능 구현 (#27)
* refractor: 일기, 답장, 발송기록 테이블 id 타입 Long으로 변경
feat: 12시에 일기 자동 발송 기록 구현(스케쥴러)

* Merge branch 'develop' into feature/mj/diary_select

# Conflicts:
#	ondam-backend/src/main/java/com/piveguyz/ondambackend/diaryRecord/command/application/controller/DiaryRecordController.java
#	ondam-backend/src/main/java/com/piveguyz/ondambackend/diaryRecord/command/application/service/DiaryRecordService.java

+++ ondam-backend/src/main/java/com/piveguyz/ondambackend/diary/command/application/controller/DiaryController.java
@@ -27,7 +27,7 @@ public class DiaryController {
         }
     }
     @DeleteMapping("/deleteDiary")
-    public ResponseEntity<String> deleteDiary(@RequestParam int id){
+    public ResponseEntity<String> deleteDiary(@RequestParam Long id){
         boolean result = diaryService.deleteDiary(id);
         if(result) {
             return ResponseEntity.status(HttpStatus.ACCEPTED).body("일기가 삭제되었습니다.");
