+++ ondam-backend/src/main/java/com/piveguyz/ondambackend/diaryRecord/command/domain/repository/DiaryRecordRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DiaryRecordRepository extends JpaRepository<DiaryRecord, Integer> {
+public interface DiaryRecordRepository extends JpaRepository<DiaryRecord, Long> {
 }
